refactor(agent): extract task-intent classifier into internal/taskintent
The CN/EN natural-language task classifier (~600 lines of keyword
tables and clause heuristics) lived inline in agent.go alongside the
run-loop orchestration it has no coupling with. It now forms the leaf
package internal/taskintent: the delivery intent type and Classify,
the task/chat heuristic, and the Goal write-budget classification.

Exported surface: Intent (+NeedsEvidence method), Classify,
NeedsEvidence, NeedsMutation, NeedsPersistentAction,
GoalNeedsWriteBudget. The redundant TaskNeedsMutation wrapper folds
into NeedsMutation; control/goal.go now calls
taskintent.GoalNeedsWriteBudget directly. Pure classifier tests move
with the package. agent.go 4667 -> 4087 lines.
